Replaced the fancy tooltip with the HTML title attribute
It's not as pretty but it helps us solve a lot of the related issues.
It's also much faster.

Added behavior so that all the commit blocks that represent the same
commit will receive the hover effect, even in other lines.

The title attribute also includes the commit's value in the specific
metric.

Fixed #52, #69, #79, #80